Add functionality: only signed in users can rate
rate-course path/page is now protected. Only signed in user can access this page. Otherwise redirect to home page.

Rate button on course page handles the scenario where user is not signed  in. It notifies user to sign in to rate the course.

Remove unused type. We can add it back if we find usage for it. But for now it seems that we are not going to use it for anything.